requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
flask-jwt-extended==4.6.0
celery==5.3.6
redis==5.0.1
flask-limiter==3.5.0
flasgger==0.9.5
python-magic==0.4.27
gunicorn==21.2.0
pytest==8.0.0
flask-socketio==5.3.6
dnspython==2.5.0
flask-mongoengine==1.0.0
requests==2.31.0
pydantic==2.6.1
python-jose==3.3.0
bcrypt==4.1.2
Werkzeug==2.3.7 